Add a method to check keys at win32 window

+++ lib/input/_win32/Win32Keyboard.cpp
@@ -6,7 +6,7 @@ using namespace Kozmic::Core::Input::Win32;
 void K_Win32Keyboard::notifyKeyUp(WPARAM t_wKey)
 {
 	for (int i = 0; i < this->m_listeners.size(); i++) {
-		std::string keyId = this->m_keyMapper->getKeyId(t_wKey);
+		std::string keyId = this->m_keyMapper->getKeyId((int) t_wKey);
 		if(keyId != "NOT_MAPPED") this->m_listeners[i]->handleKeyUp(keyId);
 	}
 }
@@ -14,7 +14,7 @@ void K_Win32Keyboard::notifyKeyUp(WPARAM t_wKey)
 void K_Win32Keyboard::notifyKeyDown(WPARAM t_wKey)
 {
 	for (int i = 0; i < this->m_listeners.size(); i++) {
-		std::string keyId = this->m_keyMapper->getKeyId(t_wKey);
+		std::string keyId = this->m_keyMapper->getKeyId((int) t_wKey);
 		if (keyId != "NOT_MAPPED") this->m_listeners[i]->handleKeyDown(keyId);
 	}
 }
